Part 1: Improvements of code and comments
