+++ include/common.h
@@ -1,0 +1,35 @@
+#ifndef COMMON_H
+#define COMMON_H
+
+// Other <.h> files
+#include "util.h"
+#include "constants.h"
+
+// All necessary libraries
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <unistd.h>
+#include <time.h>
+#include <ctype.h>
+#include <math.h>
+
+#include <sys/ipc.h>
+#include <sys/shm.h>
+#include <sys/stat.h>
+#include <sys/mman.h>
+#include <sys/select.h>
+#include <sys/wait.h>
+#include <sys/types.h> 
+#include <sys/socket.h>
+#include <netinet/in.h>
+#include <sys/time.h>
+#include <netdb.h> 
+
+#include <semaphore.h>
+#include <signal.h>
+#include <fcntl.h>
+#include <errno.h>
+// End of library declarations
+
+#endif
