+++ include/key_manager.h
@@ -8,7 +8,6 @@ void get_args(int argc, char *argv[]);
 
 int read_key_from_pipe(int pipe_fd);
 
-
 void signal_handler(int signo, siginfo_t *siginfo, void *context);
 
 /**
