+++ include/logger.h
@@ -9,6 +9,7 @@ void signal_handler(int signo, siginfo_t *siginfo, void *context);
 
 // TODO ADD COMMENTS
 void write_to_file(FILE *fp,char *msg);
+
 //TODO ADD COMMENTS
 void create_logfile_name(char *name);
