+++ include/targets.h
@@ -2,15 +2,13 @@
 #define TARGETS_H
 
 // User-defined parameters
-#define SCREEN_WIDTH 168
-#define SCREEN_HEIGHT 44
 #define MAX_TARGETS 1
 
-// Structure to represent a target
+// Structure to represent a single target
 typedef struct {
     int x;
     int y;
-} Target;
+} Target;  // Different from "targets" in constants.h
 
 void get_args(int argc, char *argv[]);
 void generateRandomCoordinates(int sectorWidth, int sectorHeight, int *x, int *y);
