+++ src/logger.c
@@ -1,21 +1,5 @@
 #include "logger.h"
-#include "constants.h"
-#include "util.h"
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <unistd.h>
-#include <time.h>
-
-#include <sys/ipc.h>
-#include <sys/shm.h>
-#include <sys/stat.h>
-#include <sys/mman.h>
-
-#include <semaphore.h>
-#include <signal.h>
-#include <fcntl.h>
-
+#include "common.h"
 
 // GLOBAL VARIABLES
 int shm_logs_fd;             // File descriptor for logs shm
