+++ src/main.c
@@ -1,14 +1,5 @@
 #include "main.h"
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <unistd.h>
-#include <time.h>
-
-#include <sys/wait.h>
-#include <sys/types.h>
-#include <signal.h>
+#include "common.h"
 
 // Serverless pipes (fd)
 int interface_km[2];
